build: Expand Markdown *before* inline TCL
This patch is a weird one. I basically just swapped two lines in the
expand proc.

This was done to fix a specific issue with running posts through #expand
twice for the amalgamation page. More generally, I think it makes the
post processing pipeline easier to reason about.

This *did* happen to break the first post because it had <??> inside a
markdown code-block which no longer works.