build+content: Add comments
This patch adds a step to the build script where comments are stripped
from posts.

These are used to explain some of the more fickle changes in the
metapost introduced in 519e330.

In the future, I think they could also be used to write down notes on
the content (rather than styling/layout) of posts.