build: Make source file structure mirror output file structure
This patch shuffles some files around such that the structure of the new
src/ directory mirrors that of _build/. This makes the repository easier
to reason about.

I also took this opportunity to move styles and images out of their
assets/ since that felt a bit arbitrary: images and css are assets but
html is firstclass?