build: Show favicons after links
This patch adds a script which injects little favicons next to links. I
used the following articles for inspiration:

- https://css-tricks.com/favicons-next-to-external-links/
- blog.jim-nielsen.com/2021/displaying-favicons-for-any-domain

An issue with the current approach is that it causes a lot of
layout-shifting during loading. This could be avoided if the
introduction of the items happened at build-time instead of client-side
at load-time.